이동방식 훅으로 분리
